s3/s3_requirements.txt
boto3==1.34.76
Flask==3.0.0
Flask-Cors==4.0.0
flasgger==0.9.7.1
python-dotenv==1.0.1
requests==2.26.0
uuid==1.30